+++ Modern JavaScript From The Beginning/Section 8: API Projects/73 - 77. Github Finder/app.js
@@ -13,7 +13,10 @@ searchInput.addEventListener(
                 .getUser(username)
                 .then(data => {
                     if (data.profile.message === 'Not Found') {
-                        // show error alert
+                        ui.showAlert(
+                            'User not found',
+                            'alert alert-danger'
+                        );
                     } else {
                         ui.showProfile(data.profile);
                     }
@@ -22,7 +25,7 @@ searchInput.addEventListener(
                     console.log(error)
                 });
         } else {
-            // clear profile
+            ui.clearProfile();
         }
     }
 );